nvme, plugins/config: migrate config-* into a config plugin
config-validate, config-show, and config-convert were flat base-level
commands, continuing what "nvme config" (the JSON-era command they
replace) already was. Move all three into a "config" plugin instead,
giving "nvme config validate/show/convert".

The old hyphenated forms keep working unchanged: nvme-cli's existing
plugin-command dispatch shorthand already treats "config-validate" as
"config validate" for any registered plugin, the same way
"registry-list" already works today.
